Telegram-Engine/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-telegram-bot
python-multipart
requests
apscheduler
passlib[bcrypt]
python-jose